Create genre filter component
